requirements.txt
flask==2.3.2
pycryptodome==3.20.0
requests==2.31.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
urllib3==1.26.18